% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{oklab_multi}
\alias{oklab_multi}
\title{Generate a multi-anchor palette via OKLAB interpolation}
\usage{
oklab_multi(colours, n = 256)
}
\arguments{
\item{colours}{Character vector of hex colour strings (at least 2).}

\item{n}{Number of colours to generate (default 256).}
}
\value{
Character vector of hex colour strings.
}
\description{
Interpolates through an arbitrary number of anchor colours in OKLAB space.
}
\examples{
oklab_multi(c("#440154", "#21918C", "#FDE725"), n = 15)
}
