% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{pheatmap_oklab_div}
\alias{pheatmap_oklab_div}
\title{Generate diverging colour vector for pheatmap}
\usage{
pheatmap_oklab_div(
  low = "#2166AC",
  mid = "#F7F7F7",
  high = "#B2182B",
  n = 256,
  space = c("oklab", "oklch")
)
}
\arguments{
\item{low}{Hex colour for the low end.}

\item{mid}{Hex colour for the midpoint.}

\item{high}{Hex colour for the high end.}

\item{n}{Number of colours (default 256).}

\item{space}{"oklab" (default) or "oklch".}
}
\value{
Character vector of hex colour strings.
}
\description{
Generate diverging colour vector for pheatmap
}
